pyproject.toml
[project]
name = "sasy-customer-demo"
version = "0.1.0"
description = "Customer-facing demo for SASY airline policy enforcement"
requires-python = ">=3.11"
dependencies = [
    "sasy>=0.4.0",
    "openai>=1.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "docstring-parser>=0.16",
]

[project.scripts]
sasy-demo = "customer_demo.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/customer_demo"]
